+++ packages/txwrapper-polkadot/src/methods/index.ts
@@ -1,2 +1,3 @@
 // Name exports to create namespaces that map to pallets
 export * as crowdloan from './crowdloan';
+export * as xcmPallet from './xcmPallet';