+++ packages/txwrapper-polkadot/src/methods/xcmPallet/reserveTransferAssets.ts
@@ -1,0 +1,40 @@
+import {
+	BaseTxInfo,
+	defineMethod,
+	OptionsWithMeta,
+	UnsignedTransaction,
+} from '@substrate/txwrapper-core';
+
+import { XcmBaseAssetArgs } from './types';
+
+export type XcmReserveTransferAssets = XcmBaseAssetArgs;
+
+/**
+ * Transfer some assets from the local chain to the sovereign account of a destination
+ * chain and forward a notification XCM.
+ *
+ * Fee payment on the destination side is made from the asset in the `assets` vector of
+ * index `fee_asset_item`. The weight limit for fees is not provided and thus is unlimited,
+ * with all fees taken as needed from the asset.
+ *
+ * @param args - Arguments specific to this method.
+ * @param info - Information required to construct the transaction.
+ * @param options - Registry and metadata used for constructing the method.
+ */
+export function reserveTransferAssets(
+	args: XcmReserveTransferAssets,
+	info: BaseTxInfo,
+	options: OptionsWithMeta
+): UnsignedTransaction {
+	return defineMethod(
+		{
+			method: {
+				args,
+				name: 'reserveTransferAssets',
+				pallet: 'xcmPallet',
+			},
+			...info,
+		},
+		options
+	);
+}