+++ packages/txwrapper-template/CHANGELOG.md
@@ -3,6 +3,62 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.2.1](http://%3Clink%20to%20the%20github%20repo%20this%20package%20lives%20in%3E/compare/v3.2.0...v3.2.1) (2022-08-30)
+
+**Note:** Version bump only for package @substrate/txwrapper-template
+
+
+
+
+
+# [3.2.0](http://%3Clink%20to%20the%20github%20repo%20this%20package%20lives%20in%3E/compare/v3.1.7...v3.2.0) (2022-08-16)
+
+**Note:** Version bump only for package @substrate/txwrapper-template
+
+
+
+
+
+## [3.1.7](http://%3Clink%20to%20the%20github%20repo%20this%20package%20lives%20in%3E/compare/v3.1.6...v3.1.7) (2022-07-26)
+
+**Note:** Version bump only for package @substrate/txwrapper-template
+
+
+
+
+
+## [3.1.6](http://%3Clink%20to%20the%20github%20repo%20this%20package%20lives%20in%3E/compare/v3.1.5...v3.1.6) (2022-07-12)
+
+**Note:** Version bump only for package @substrate/txwrapper-template
+
+
+
+
+
+## [3.1.5](http://%3Clink%20to%20the%20github%20repo%20this%20package%20lives%20in%3E/compare/v3.1.4...v3.1.5) (2022-07-05)
+
+**Note:** Version bump only for package @substrate/txwrapper-template
+
+
+
+
+
+## [3.1.4](http://%3Clink%20to%20the%20github%20repo%20this%20package%20lives%20in%3E/compare/v3.1.3...v3.1.4) (2022-06-27)
+
+**Note:** Version bump only for package @substrate/txwrapper-template
+
+
+
+
+
+## [3.1.3](http://%3Clink%20to%20the%20github%20repo%20this%20package%20lives%20in%3E/compare/v3.1.2...v3.1.3) (2022-06-20)
+
+**Note:** Version bump only for package @substrate/txwrapper-template
+
+
+
+
+
 ## [3.1.2](http://%3Clink%20to%20the%20github%20repo%20this%20package%20lives%20in%3E/compare/v3.1.1...v3.1.2) (2022-06-15)
 
 **Note:** Version bump only for package @substrate/txwrapper-template